refactor tests

+++ src/models/llama.rs
@@ -596,7 +596,7 @@ mod tests {
                 slot_mapping: Tensor::new(&[tokens.len() as i64 - 1], &device)?,
                 decoding_metadata: Some(FlashAttentionDecodingMetadata {
                     block_tables: Some(
-                        Tensor::arange(0, num_blocks, &device)?.reshape((1, num_blocks))?,
+                        Tensor::arange(0, num_blocks, &device)?.reshape((1, num_blocks as usize))?,
                     ),
                     max_decoding_sequence_length: tokens.len(),
                     sequence_lengths: Some(Tensor::new(&[tokens.len() as u32], &device)?),
